Various CTF Challenges
STEM-Forensics-100-TMNT: Teenage Mutant Ninja Turtles Security
challenge. The user has to use strings to read hidden ASCII in various
images. Then submit those to the web application to get the flag.

STEM-Forensics-200-DTMF: Dual-tone multi-frequency signaling (phone
number noise). The user has to listen to a tone when translates to hex
values. This is the flag.

STEM-Networking-100-MatrixUDP: The user has to simply send a UDP packet
to the server, the server responds with a flag.

WCSC-Networking-Weekly12-PortalSong: The user connects to a service on
the service which prompts the first line of the "Portal" song. The user
has to play back the next line, the server follows up. The connection is
closed after a few seconds so to get to the flag the entire process has
to be scripted.

STEM-Web-100-InitechEmail: The user has to edit the form post parameters
to alter the sender of the e-mail. This is essentially spoofing another
user's e-mail address. Then they receive an e-mail with the flag.